fix: prepare plugin for NativeScript 6.0.0 release
In NativeScript 6.0.0 there's a major refactoring in CLI that will require changes in most of the plugin hooks.
The major changes are the hookArgs injected by CLI for each hook - they are changed, so some of the properties used from them, should be taken from different property.
The other major change is the injected data - in the function used as hook, you can use any service registered in CLI's bootstrap and CLI will pass it to the hook.
Several of the services are now renamed or deleted, so hooks using them should be migrated. Such service is platformsData.
The last major change is that some logger methods are deleted (they've been deprecated).

The current PR makes the hooks of this plugin compatible with the both 6.0.0 and old releases of NativeScript.